fix: Router conflict with ErrorTracker
Moved parse_options/2 to PhoenixAnalytics.Web.RouterHelpers module
